move factories for infrastructure models into infrastructure factories
